Change boot.sh commenting to test /data/etc/.boot_to_xbmc

+++ hack/bootxbmc.sh
@@ -1,9 +1,12 @@
 #!/bin/sh
-/etc/rc3.d/U94boxeehal stop
-/etc/rc3.d/U99boxee stop
-killall BoxeeHal
-killall BoxeeLauncher
-killall Boxee
-killall run_boxee.sh
-mount -o bind /data/hack/xbmc.sh /opt/boxee/BoxeeLauncher
-break
+if [ /data/etc/.boot_to_xbmc == "1" ]
+then
+	/etc/rc3.d/U94boxeehal stop
+	/etc/rc3.d/U99boxee stop
+	killall BoxeeHal
+	killall BoxeeLauncher
+	killall Boxee
+	killall run_boxee.sh
+	mount -o bind /data/hack/xbmc.sh /opt/boxee/BoxeeLauncher
+	break
+fi
